Moved FileDrop to /component
